New Objects for executing Reports with RequestPages from External
This point refers to Issue 546 of Microsoft/AL. In it the request was introduced to make it possible to read out the options and data items of the request page via the object metadata table. Unfortunately, the NAV standard does not allow to do this. And retrieval of the Request XML document is only possible when the client is started and the RequestPage is displayed to the user. Furthermore, a lot of important information is missing to prepare this XML document accordingly. This customization is required to start a report and request page from an external Web service, and to obtain the required request parameters in order to be able to build up the XML required by NAV. The generated XML has the same structure as the query XML file, which can be passed to the REPORT. SAVE function. Only the missing options such as data type and additional information have been added.